feat(client): make third-party Sign in with Seed usable from npm
Third-party sites could not integrate with Seed from a published package.
The auth protocol lived in @shm/shared (unpublished), the browser session
logic lived inside the web app (not a package at all), and the one package
that is on npm was broken for consumers.

Move the protocol into @seed-hypermedia/client:

- blobs, cbor and hmauth move from @shm/shared. They only depend on noble,
  multiformats and cborg, all already client dependencies, and hmauth
  already imported @seed-hypermedia/client/base64 — so this removes a
  dependency inversion rather than adding one. @shm/shared keeps thin
  re-export shims, so all 46 existing import sites are untouched.
- Add src/auth.ts, a framework-free port of the web app's auth-session
  module: startAuth, handleCallback, getSession, signWithSession,
  clearSession, plus createSessionSigner to adapt a session into an
  HMSigner for the existing blob builders. Session storage is pluggable
  and defaults to IndexedDB, which is required to persist a
  non-extractable CryptoKey across the redirect.

Fix the published package, which could not be imported at all:

- npm publish does not apply publishConfig field overrides (that is a pnpm
  feature), so 0.0.68 shipped exports pointing at ./src/index.ts while
  files only included dist/. A bare import failed with
  ERR_UNSUPPORTED_NODE_MODULES_TYPE_STRIPPING. Publish with pnpm instead.
- publishConfig pointed at ./dist/index.js, but tsup emits .mjs, so the
  paths would not have resolved even once applied.
- tsup only built 5 entrypoints while tsc emitted .d.ts for every module,
  so the ./* subpath export could resolve types with no runtime file.
  Build every module.
- Narrow the zod peer range to ^3.20.0. zod 4 removed
  ZodDiscriminatedUnion.optionsMap, which HMBlockSchema reads at import
  time, so a bare import threw under zod 4.

Add the tarball install-and-import check the CLI job already had and the
client job lacked, which is why the broken manifest shipped unnoticed.
